requirements_streamlit.txt
streamlit
plotly
pandas
psycopg2-binary
